add tweets_controller.rb and profile.html.slim
