+++ app/assets/stylesheets/details.scss
@@ -24,6 +24,11 @@
   padding-right: 5px;
 }
 
+.space .link {
+  text-decoration: none;
+  color: white;
+}
+
 .edit {
   text-decoration: none;
   color: white;
