+++ config/routes.rb
@@ -21,6 +21,10 @@ Rails.application.routes.draw do
     resources :comments, only: %i[create destroy edit]
   end
 
+  resources :tweets, only: %i[new create]
+
+  get '/profile' => 'users#profile'
+
   resources :home, only: %i[index new create]
 
   mount LetterOpenerWeb::Engine, at: 'letter_opener' if Rails.env.development?
